requirements.txt
# MagicLight v2.0 — Requirements

# Google APIs
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0

# Browser Automation
playwright==1.44.0

# Video Processing (FFmpeg must be installed separately)
# No Python package needed — uses subprocess

# Utilities
python-dotenv==1.0.1
